08_tools/tools/pyproject.toml
[project]
name = "tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.19",
    "playwright>=1.52.0",
    "python-dotenv>=1.1.0",
]
